init: u-boot-2019.04

+++ u-boot-sp/include/usb/dwc2_udc.h
@@ -1,9 +1,8 @@
+/* SPDX-License-Identifier: GPL-2.0+ */
 /*
  * drivers/usb/gadget/dwc2_udc.h
  * Designware DWC2 on-chip full/high speed USB device controllers
  * Copyright (C) 2005 for Samsung Electronics
- *
- * SPDX-License-Identifier:	GPL-2.0+
  */
 
 #ifndef __DWC2_USB_GADGET
@@ -15,7 +14,7 @@ struct dwc2_plat_otg_data {
 	void		*priv;
 	int		phy_of_node;
 	int		(*phy_control)(int on);
-	unsigned int	regs_phy;
+	uintptr_t	regs_phy;
 	uintptr_t	regs_otg;
 	unsigned int    usb_phy_ctrl;
 	unsigned int    usb_flags;
